Beveiliging
Hele hoop kleine dingen veranderd in de beveiliging van de website.

+++ PHP/delete.php
@@ -1,16 +1,20 @@
 <?php
-require_once("../Includes/connector.php");
+if(isset($_GET['id'])) {
+    require_once("../Includes/connector.php");
 
-$dataTable = $_GET['dataTable'];
+    $dataTable = $_GET['dataTable'];
 
-$sql = "DELETE FROM $dataTable WHERE ID =:ID";
+    $sql = "DELETE FROM $dataTable WHERE ID =:ID";
 
-$stmt = $connect->prepare($sql);
-$stmt->bindParam(":ID", $_GET['id']);
-$stmt->execute();
+    $stmt = $connect->prepare($sql);
+    $stmt->bindParam(":ID", $_GET['id']);
+    $stmt->execute();
 
-if ($dataTable === "reserveringen" ||  $dataTable === "contact") {
-    header ("Location: ../AdminPages/adminReserveringen-contact.php");
+    if ($dataTable === "reserveringen" ||  $dataTable === "contact") {
+        header ("Location: ../AdminPages/adminReserveringen-contact.php");
+    } else {
+        header ("Location: ../AdminPages/changeMenu.php");
+    }
 } else {
     header ("Location: ../AdminPages/changeMenu.php");
 }
